Set timeout on initial connection to the resolver
The existing timeout in _try_query only takes effect if a TCP connection
has already been established. If a connection has not yet been
established, and the resolver is not accepting new connections, then
this server could end up waiting indefinitely for a connection to be
established.  Instead, we wrap self.loop.create_connection in
asyncio.wait_for so that we are able to set a timeout.